[amount] Add new amount package
**Context**:
Currently, amount related logics are implemented as functions in util
packages. All the other places uses `currency` package for amounts,
which leads to many inconsistencies. Also, this approach makes
formatting harder as formatting in `currency` package is still WIP.

This change adds a custom type `RawAmountVal`, which is intended to be
used as raw currency amount type throughout the program. This also
decouples the underlying currency with just the raw amount.

**Changes**:
- [amount] add amount package
- [amount] add related unit testings