[profile] fix: category operation take relative path
Currently add/find category matches the first token with the
parent category node. This caused the fact that every traversal
needs to use a full path starting with the name of the parent category
node. A better approach is to ignore the name of the category node where
add/find operation are called.

**Changes**:
- changed add/find category
- corresponding changes in unit tests