[profile/category] add visitor functions
**Background**:
While developing the stat command, I found it will be useful
we have generic visitor-like functions for the category tree.
We can use it to calculate sum, or filter out some categories.

Another good-to-have thing is path prefix in each node. Then
we can store those path prefixes in SQL as opposed to just
category name. This way, when we list the records out, we can
have cateogry path which makes the list clearer

**Changes**:
Making amends to `Category` type
- Add `Visit` function that takes a selector (filter) and an accumulator
- Add corresponding tests
- Add path prefixes to each Category node